+++ next.config.mjs
@@ -158,10 +158,13 @@ const withMDX = createMDX({
     },
 });
 
-export default withSentryConfig(withMDX(nextConfig), {
-    silent: true,
-    authToken: process.env.SENTRY_AUTH_TOKEN,
-    org: process.env.SENTRY_ORG,
-    project: process.env.SENTRY_PROJECT,
-    sentryUrl: process.env.SENTRY_URL,
-});
+const finalConfig = withMDX(nextConfig);
+export default process.env.SENTRY_AUTH_TOKEN
+    ? withSentryConfig(finalConfig, {
+          silent: true,
+          authToken: process.env.SENTRY_AUTH_TOKEN,
+          org: process.env.SENTRY_ORG,
+          project: process.env.SENTRY_PROJECT,
+          sentryUrl: process.env.SENTRY_URL,
+      })
+    : finalConfig;