+++ src/app/layout.jsx
@@ -221,7 +221,7 @@ export default async function RootLayout({ children }) {
                     <Footer />
                     <BottomNav />
                 </LiveDataProvider>
-                {isProduction ?
+                {isProduction && process.env.UMAMI_SITE_ID ?
                     <Script
                         nonce={nonce}
                         // src="https://umami.drunik.be/script.js"