+++ src/auth.js
@@ -15,34 +15,36 @@ import { betterAuth } from 'better-auth';
 import { prismaAdapter } from 'better-auth/adapters/prisma';
 import db from '@/db/db';
 
-export const auth = betterAuth({
-    secret: process.env.BETTER_AUTH_SECRET,
-    baseURL: process.env.BETTER_AUTH_URL,
-    database: prismaAdapter(db, {
-        provider: 'postgresql',
-    }),
-    account: {
-        accountLinking: {
-            trustedProviders: ['discord', 'github'],
-        },
-    },
-    socialProviders: {
-        discord: {
-            clientId: process.env.AUTH_DISCORD_ID,
-            clientSecret: process.env.AUTH_DISCORD_SECRET,
-        },
-        github: {
-            clientId: process.env.AUTH_GITHUB_ID,
-            clientSecret: process.env.AUTH_GITHUB_SECRET,
-        },
-    },
-    user: {
-        additionalFields: {
-            role: {
-                type: 'string',
-                defaultValue: 'user',
-                input: false,
-            },
-        },
-    },
-});
+export const auth = process.env.BETTER_AUTH_SECRET
+    ? betterAuth({
+          secret: process.env.BETTER_AUTH_SECRET,
+          baseURL: process.env.BETTER_AUTH_URL,
+          database: prismaAdapter(db, {
+              provider: 'postgresql',
+          }),
+          account: {
+              accountLinking: {
+                  trustedProviders: ['discord', 'github'],
+              },
+          },
+          socialProviders: {
+              discord: {
+                  clientId: process.env.AUTH_DISCORD_ID,
+                  clientSecret: process.env.AUTH_DISCORD_SECRET,
+              },
+              github: {
+                  clientId: process.env.AUTH_GITHUB_ID,
+                  clientSecret: process.env.AUTH_GITHUB_SECRET,
+              },
+          },
+          user: {
+              additionalFields: {
+                  role: {
+                      type: 'string',
+                      defaultValue: 'user',
+                      input: false,
+                  },
+              },
+          },
+      })
+    : null;